Rename build-rows and jsonl-stream to camelCase
Matches the repo-wide convention for service module filenames (e.g.
referenceUtils.ts, vocabularyService.ts).
